refactor: extract auth HTTP routes into auth_routes feature-module package (#826)
Follows the quizzes reference implementation (#825) to migrate the next
domain in the tracking issue: /auth/* (Keycloak login/register/callback/
logout), /api/auth/* (config, metadata, login, logout, OTP request/login),
and the authenticated /api/auth/me.

Unlike quizzes, business logic stays in the existing news_dashboard.auth
module rather than moving into a new service.py: auth.py is imported across
~60 files (require_auth, require_admin, session/OTP/Keycloak helpers), so
renaming it into a package would touch far more than this domain's routes.
auth_routes/ is a thin router layer on top of it, mirroring the existing
greader.py split between a public router and an authenticated one.

- auth_routes/router.py: public_router (mounted on main's public_router,
  no auth) plus router (mounted on main's authenticated api router) for
  /api/auth/me.
- auth_routes/models.py: LoginRequest, OTPRequestPayload, OTPLoginPayload.
- No route paths or auth semantics changed.
- Updated test_main_oauth.py patch targets to the new module location.
- Added test_auth_routes_package.py asserting the package imports and every
  route stays mounted (mirrors test_quizzes_package.py).

Tests: pytest tests/test_auth.py tests/test_auth_extra.py
tests/test_main_oauth.py tests/test_auth_routes_package.py
tests/test_quizzes_package.py -> 115 passed
ruff check . -> all checks passed
mypy news_dashboard -> no issues in 94 source files
